+++ commons/ava_services/gecko_service.go
@@ -1,17 +1,19 @@
 package ava_services
 
 import (
-	"github.com/kurtosis-tech/kurtosis/commons/services"
+	"github.com/docker/go-connections/nat"
 )
 
 type GeckoService struct {
 	ipAddr string
+	stakingPort nat.Port
+	jsonRpcPort nat.Port
 }
 
-func (g GeckoService) GetStakingSocket() services.ServiceSocket {
-	return *services.NewServiceSocket(g.ipAddr, stakingPort)
+func (g GeckoService) GetStakingSocket() ServiceSocket {
+	return *NewServiceSocket(g.ipAddr, g.stakingPort)
 }
 
-func (g GeckoService) GetJsonRpcSocket() services.ServiceSocket {
-	return *services.NewServiceSocket(g.ipAddr, httpPort)
+func (g GeckoService) GetJsonRpcSocket() ServiceSocket {
+	return *NewServiceSocket(g.ipAddr, g.jsonRpcPort)
 }