+++ commons/ava_services/gecko_service_availability_checker_core.go
@@ -37,7 +37,7 @@ func (g GeckoServiceAvailabilityCheckerCore) IsServiceUp(toCheck services.Servic
 }
 
 func (g GeckoServiceAvailabilityCheckerCore) GetTimeout() time.Duration {
-	return 30 * time.Second
+	return 90 * time.Second
 }