+++ controller/main.go
@@ -35,10 +35,16 @@ func main() {
 		"Comma-separated list of specific tests to run (leave empty or omit to run all tests)",
 	)
 
-	testImageNameArg := flag.String(
-		"test-image-name",
+	geckoImageNameArg := flag.String(
+		"gecko-image-name",
 		"",
-		"Name of Docker image of the service being tested",
+		"Name of Docker image of the Gecko version being tested",
+	)
+
+	chitSpammerImageNameArg := flag.String(
+		"chit-spammer-image-name",
+		"",
+		"The name of a pre-built Gecko image, either on the local Docker engine or in Docker Hub",
 	)
 
 	dockerNetworkArg := flag.String(
@@ -88,8 +94,14 @@ func main() {
 		*subnetMaskArg,
 		*gatewayIpArg,
 		*testControllerIpArg,
-		*testImageNameArg)
+		*geckoImageNameArg)
 
+
+	logrus.Debugf("Chit spammer image name: %s", *chitSpammerImageNameArg)
+	testSuite := ava_testsuite.AvaTestSuite{
+		ChitSpammerImageName: *chitSpammerImageNameArg,
+		NormalImageName:      *geckoImageNameArg,
+	}
 	controller := controller.NewTestController(
 		*testVolumeArg,
 		*testVolumeMountpointArg,
@@ -97,11 +109,11 @@ func main() {
 		*subnetMaskArg,
 		*gatewayIpArg,
 		*testControllerIpArg,
-		ava_testsuite.AvaTestSuite{},
-		*testImageNameArg)
+		testSuite,
+		*testNameArg)
 
 	logrus.Infof("Running test '%v'...", *testNameArg)
-	setupErr, testErr := controller.RunTest(*testNameArg)
+	setupErr, testErr := controller.RunTest()
 	if setupErr != nil {
 		logrus.Errorf("Test %v encountered an error during setup (test did not run):", *testNameArg)
 		fmt.Fprintln(logrus.StandardLogger().Out, setupErr)